Crate map_split

Structs

HashMapA
A wrapper around a HashMap which provides access to the A portion of a Splittable value type

HashMapB
A wrapper around a HashMap which provides access to the B portion of a Splittable value type

HashMapC
A wrapper around a HashMap which provides access to the C portion of a Splittable3 value type

HashMapD
A wrapper around a HashMap which provides access to the D portion of a Splittable4 value type

HashMapE
A wrapper around a HashMap which provides access to the E portion of a Splittable5 value type

HashMapF
A wrapper around a HashMap which provides access to the F portion of a Splittable6 value type

HashMapG
A wrapper around a HashMap which provides access to the G portion of a Splittable7 value type

HashMapH
A wrapper around a HashMap which provides access to the H portion of a Splittable8 value type

HashMapI
A wrapper around a HashMap which provides access to the I portion of a Splittable9 value type

HashMapJ
A wrapper around a HashMap which provides access to the J portion of a Splittable10 value type

HashMapK
A wrapper around a HashMap which provides access to the K portion of a Splittable11 value type

HashMapL
A wrapper around a HashMap which provides access to the L portion of a Splittable12 value type

HashMapM
A wrapper around a HashMap which provides access to the M portion of a Splittable13 value type

HashMapN
A wrapper around a HashMap which provides access to the N portion of a Splittable14 value type

HashMapO
A wrapper around a HashMap which provides access to the O portion of a Splittable15 value type

HashMapP
A wrapper around a HashMap which provides access to the P portion of a Splittable16 value type

HashMapQ
A wrapper around a HashMap which provides access to the Q portion of a Splittable17 value type

HashMapR
A wrapper around a HashMap which provides access to the R portion of a Splittable18 value type

HashMapS
A wrapper around a HashMap which provides access to the S portion of a Splittable19 value type

HashMapT
A wrapper around a HashMap which provides access to the T portion of a Splittable20 value type

HashMapU
A wrapper around a HashMap which provides access to the U portion of a Splittable21 value type

HashMapV
A wrapper around a HashMap which provides access to the V portion of a Splittable22 value type

HashMapW
A wrapper around a HashMap which provides access to the W portion of a Splittable23 value type

HashMapX
A wrapper around a HashMap which provides access to the X portion of a Splittable24 value type

HashMapY
A wrapper around a HashMap which provides access to the Y portion of a Splittable25 value type

HashMapZ
A wrapper around a HashMap which provides access to the Z portion of a Splittable26 value type

Traits

Splittable
A type which can be split into disjoint references. SplitType allows a type to be split multiple

Splittable10
A type which can be split into disjoint references. SplitType allows a type to be split multiple

Splittable11
A type which can be split into disjoint references. SplitType allows a type to be split multiple

Splittable12
A type which can be split into disjoint references. SplitType allows a type to be split multiple

Splittable13
A type which can be split into disjoint references. SplitType allows a type to be split multiple

Splittable14
A type which can be split into disjoint references. SplitType allows a type to be split multiple

Splittable15
A type which can be split into disjoint references. SplitType allows a type to be split multiple

Splittable16
A type which can be split into disjoint references. SplitType allows a type to be split multiple

Splittable17
A type which can be split into disjoint references. SplitType allows a type to be split multiple

Splittable18
A type which can be split into disjoint references. SplitType allows a type to be split multiple

Splittable19
A type which can be split into disjoint references. SplitType allows a type to be split multiple

Splittable20
A type which can be split into disjoint references. SplitType allows a type to be split multiple

Splittable21
A type which can be split into disjoint references. SplitType allows a type to be split multiple

Splittable22
A type which can be split into disjoint references. SplitType allows a type to be split multiple

Splittable23
A type which can be split into disjoint references. SplitType allows a type to be split multiple

Splittable24
A type which can be split into disjoint references. SplitType allows a type to be split multiple

Splittable25
A type which can be split into disjoint references. SplitType allows a type to be split multiple

Splittable26
A type which can be split into disjoint references. SplitType allows a type to be split multiple

Splittable3
A type which can be split into disjoint references. SplitType allows a type to be split multiple

Splittable4
A type which can be split into disjoint references. SplitType allows a type to be split multiple

Splittable5
A type which can be split into disjoint references. SplitType allows a type to be split multiple

Splittable6
A type which can be split into disjoint references. SplitType allows a type to be split multiple

Splittable7
A type which can be split into disjoint references. SplitType allows a type to be split multiple

Splittable8
A type which can be split into disjoint references. SplitType allows a type to be split multiple

Splittable9
A type which can be split into disjoint references. SplitType allows a type to be split multiple

Functions

split
Splits a HashMap into two disjoint hashmap references, able to access the split parts of the stored Splittable values independently.

split10
Splits a HashMap into 10 disjoint hashmap references, able to access the split parts of the stored Splittable10 values independently.

split11
Splits a HashMap into 11 disjoint hashmap references, able to access the split parts of the stored Splittable11 values independently.

split12
Splits a HashMap into 12 disjoint hashmap references, able to access the split parts of the stored Splittable12 values independently.

split13
Splits a HashMap into 13 disjoint hashmap references, able to access the split parts of the stored Splittable13 values independently.

split14
Splits a HashMap into 14 disjoint hashmap references, able to access the split parts of the stored Splittable14 values independently.

split15
Splits a HashMap into 15 disjoint hashmap references, able to access the split parts of the stored Splittable15 values independently.

split16
Splits a HashMap into 16 disjoint hashmap references, able to access the split parts of the stored Splittable16 values independently.

split17
Splits a HashMap into 17 disjoint hashmap references, able to access the split parts of the stored Splittable17 values independently.

split18
Splits a HashMap into 18 disjoint hashmap references, able to access the split parts of the stored Splittable18 values independently.

split19
Splits a HashMap into 19 disjoint hashmap references, able to access the split parts of the stored Splittable19 values independently.

split20
Splits a HashMap into 20 disjoint hashmap references, able to access the split parts of the stored Splittable20 values independently.

split21
Splits a HashMap into 21 disjoint hashmap references, able to access the split parts of the stored Splittable21 values independently.

split22
Splits a HashMap into 22 disjoint hashmap references, able to access the split parts of the stored Splittable22 values independently.

split23
Splits a HashMap into 23 disjoint hashmap references, able to access the split parts of the stored Splittable23 values independently.

split24
Splits a HashMap into 24 disjoint hashmap references, able to access the split parts of the stored Splittable24 values independently.

split25
Splits a HashMap into 25 disjoint hashmap references, able to access the split parts of the stored Splittable25 values independently.

split26
Splits a HashMap into 26 disjoint hashmap references, able to access the split parts of the stored Splittable26 values independently.

split3
Splits a HashMap into 3 disjoint hashmap references, able to access the split parts of the stored Splittable3 values independently.

split4
Splits a HashMap into 4 disjoint hashmap references, able to access the split parts of the stored Splittable4 values independently.

split5
Splits a HashMap into 5 disjoint hashmap references, able to access the split parts of the stored Splittable5 values independently.

split6
Splits a HashMap into 6 disjoint hashmap references, able to access the split parts of the stored Splittable6 values independently.

split7
Splits a HashMap into 7 disjoint hashmap references, able to access the split parts of the stored Splittable7 values independently.

split8
Splits a HashMap into 8 disjoint hashmap references, able to access the split parts of the stored Splittable8 values independently.

split9
Splits a HashMap into 9 disjoint hashmap references, able to access the split parts of the stored Splittable9 values independently.