Struct map_split::HashMapH

pub struct HashMapH<'a, K: 'a, V: 'a, S: 'a, SplitType>(_, _) where K: Eq + Hash, S: BuildHasher;

A wrapper around a HashMap which provides access to the H portion of a Splittable8 value type

Methods

impl<'a, K: 'a, V: 'a, S, SplitType> HashMapH<'a, K, V, S, SplitType> where K: Eq + Hash, S: BuildHasher

fn iter<'b>(&'b self) -> Map<Iter<'b, K, V>, fn((&'b K, &'b V)) -> (&'b K, V::H)> where V: Splittable8<'b, SplitType>

An iterator visiting all key-value pairs in arbitrary order. Iterator element type is (&'b K, &'b V::H).

fn iter_mut<'b>(&'b mut self) -> Map<IterMut<'b, K, V>, fn((&'b K, &'b mut V)) -> (&'b K, V::MutH)> where V: Splittable8<'b, SplitType>

An iterator visiting all key-value pairs in arbitrary order. Iterator element type is (&'b K, &'b mut V::H).

fn capacity(&self) -> usize

Returns the number of elements the map can hold without reallocating.

This number is a lower bound; the HashMap might be able to hold more, but is guaranteed to be able to hold at least this many.

fn len(&self) -> usize

Returns the number of elements in the map.

fn is_empty(&self) -> bool

Returns true if the map contains no elements.

fn get<'b, Q: ?Sized>(&'b self, k: &Q) -> Option<V::H> where Q: Hash + Eq, K: Borrow<Q>, V: Splittable8<'b, SplitType>

Returns a reference to the value corresponding to the key.

The key may be any borrowed form of the map's key type, but Hash and Eq on the borrowed form must match those for the key type.

fn get_mut<'b, Q: ?Sized>(&'b mut self, k: &Q) -> Option<V::MutH> where Q: Hash + Eq, K: Borrow<Q>, V: Splittable8<'b, SplitType>

Returns a mutable reference to the value corresponding to the key.

The key may be any borrowed form of the map's key type, but Hash and Eq on the borrowed form must match those for the key type.